requirements.txt
pytorch-lightning >= 1.0.0rc2
torch >= 1.3.0
torchvision >= 0.6.0
timm==0.3.4
PyYAML==5.4.1
pytest==6.2.2
pandas==1.2.1
easydict==1.9
torchmetrics==0.3.0rc0